requirements.txt
python-telegram-bot==21.3
requests==2.32.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0